+++ src/core/techniques/file_based/fb_injector.py
@@ -16,7 +16,7 @@ For more see the file 'readme/COPYING' for copying permission.
 from src.core.controller import injector
 
 """
-The "file-based" technique on semiblind OS command injection.
+The "file-based" technique on blind OS command injection.
 """
 
 """